Moving card out of component
