+++ src/components/AverageStats/AverageStats.tsx
@@ -1,28 +1,4 @@
-type AverageCardProps = {
-  title: string;
-  description: string;
-  dataText: string;
-  dataDesc: string;
-};
-
-const AverageCard = ({
-  title,
-  description,
-  dataText,
-  dataDesc,
-}: AverageCardProps) => (
-  <div className="average-card">
-    <div className="average-card-title">
-      <p className="average-card-header text-preset-6">
-        {title} <p className="average-card-desc text-preset-7">{description}</p>
-      </p>
-    </div>
-    <div className="average-data">
-      <p className="average-data-text text-preset-4">{dataText}</p>
-      <p className="average-data-desc text-preset-7">{dataDesc}</p>
-    </div>
-  </div>
-);
+import AverageCard from "./AverageCard";
 
 const AverageStats = () => {
   return (
